refactor: remove utils folder and move task-utils to lib, and update imports
